+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### New
 
 - Field `content_type` and `content_encoding` added to the `amqp_0_9` output.
+- Batching fields added to the `hdfs` output.
 
 ## 3.36.0 - 2020-12-24